+++ src/components/reaction/index.tsx
@@ -4,7 +4,8 @@ import { use, useCallback, useRef, useState } from 'react';
 
 import Button from '@/components/reaction/button';
 import { useAction } from '@/contexts/action/use-action';
-import { postReaction, type Reaction } from '@/lib/supabase/client/reaction';
+import { postReaction } from '@/lib/supabase/client/reaction';
+import type { Reaction } from '@/lib/supabase/server/reaction';
 import { getItem } from '@/utils/local-storage';
 import { launch } from '@/utils/particle';
